+++ src/main/kotlin/dev/bendik/domain/ParseResult.kt
@@ -1,4 +1,4 @@
 package dev.bendik.domain
 
-data class ParseResult(val labels: Map<String, Int>,
+data class ParseResult(val labels: Map<String, Long>,
                        val instructions: List<Instruction>)
